# backend/docs/swagger.yaml
basePath: /
definitions:
...
  utils.ErrorDetail:
    properties:
      field:
        type: string
      message:
        type: string
    type: object
  utils.ErrorResponse:
    properties:
      error:
        properties:
          code:
            type: string
          details:
            items:
              $ref: '#/definitions/utils.ErrorDetail'
            type: array
          message:
            type: string
          request_id:
            type: string
          timestamp:
            type: string
        type: object
    type: object
...
host: localhost:8080
info:
  contact:
    name: API Support
    url: https://example.com/support
  description: API for Reelcut — clip creation from long-form video (auth, projects,
    videos, transcriptions, analysis, clips, templates, jobs).
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://example.com/terms
  title: Reelcut API
  version: "1.0"
paths:
...
  /api/v1/users/me:
    delete:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete current user account
      tags:
      - users
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get current user profile
      tags:
      - users
    put:
      consumes:
      - application/json
      parameters:
      - description: full_name, avatar_url
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update current user profile
      tags:
      - users
  /api/v1/users/me/usage:
    get:
      parameters:
      - description: Page
        in: query
        name: page
        type: integer
      - description: Per page
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get current user usage stats
      tags:
      - users
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"